frame rate independance.  uses float for internal position updates (otherwise position wouldn't update at high framerates).
I added setPosf and other similar funcs for the update code to use,
while the rest of the code still uses the integer interface (to keep it
all on a grid).


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@670 84b32ba4-53c3-423c-be69-77cca6335494

+++ examples/lr/player.cpp
@@ -19,9 +19,9 @@ void player::initTexture(){
 }
 
 void player::draw(){
-    int realh = height;
-	int gameh = (height+29)/30;
-	int p = pos/10;
+    int realh = (int)height;
+	int gameh = (int)((height+29.0f)/30.0f);
+	int p = (int)(pos/10.0f);
 	
 //	glColor3f(1.0f, 0.0f, 0.0f);
     glEnable(GL_TEXTURE_2D);
